Update convert_ASCIIZ_string_to_32_bit_integer
A function that converts a string of ASCII digits into a 32-bit integer. The function will receive as an argument the starting address of the string and must return a 32-bit integer containing the integer value of the string. The string is an ASCIIZ string, i.e., ends with the null character (ASCII code 0). Do not need to check for errors in the string, i.e.,  assume the string contains only characters '0' through '9' (i.e., their corresponding ASCII codes), and not represent a negative number or a non-decimal value or too large a number. For example, a_to_i called with the argument "00012345" will return the integer 12345.atoi()